1/n improve haneul-json-rpc error codes and handling
## Description 
Most of the errors we return in `haneul-json-rpc` are `anyhow::Error`s that
then get converted to `RpcError`. Or, if we keep things in
`haneul_json_rpc::Error`, there is a catch-all `From` implementation that
converts all variants to `CallError::Failed`.

In this PR, the `From` implementation is expanded to map some variants
to `RpcError::Call(CallError::InvalidParams` instead, so that we can
return different error codes rather than defaulting to `-32000` for
everything.

This will help w/ differentiating server from client errors, and we can
make use of this for alerting on error logs.
Start with `UserInputErrors`, which now map to error code `-32602`. Also
log error code if available.

In terms of performance, there shouldn't be any impact since we do the
same number of conversions. Only, instead of `JoinError` ->
`anyhow::Error` -> `RpcError`, or `anyhow::Error` -> `anyhow::Error` ->
`RpcError`, we now do `error from inner call` -> `haneul_json_rpc::Error`
-> `RpcError`.

```
curl --location '127.0.0.1:9000' \
--header 'Content-Type: application/json' \
--data '{
  "jsonrpc": "2.0",
  "id": 1,
  "method": "haneul_getTransactionBlock",
  "params": [
    [
      "DwRWCvrkfauz95wTrfJxi3VSWdMZWCEtdPCiaZhcanZh"
    ],
    {
      "showInput": true,
      "showRawInput": false,
      "showEffects": true,
      "showEvents": true,
      "showObjectChanges": false,
      "showBalanceChanges": false
    }
  ]
}'
{"jsonrpc":"2.0","error":{"code":-32602,"message":"invalid type: sequence, expected a string at line 2 column 4"},"id":1}%       
```

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes