[1/x] move some test util functions to wallet context
## Description 

First step towards moving functions in `test-utils/transaction` to
places where they belong better. This PR deals with some of functions
that take in a `WalletContext`. Some of them are moved to
`wallet_context.rs` and a few are moved to the caller site. More wallet
context functions will be moved in following PRs.

## Test Plan 

It's all tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes