Replay from dead/forked node: Introduce `NodeStateDump`
## Description 

This introduces `NodeStateDump` which is the set of objects and info
required to reproduce a fork in replay.
The created state is a few MB, so not too concerning. Json was so the
data is human readable.

Most of the logic in this PR is in getting the replay system to load
from state dump.

Sample config
```
state-debug-dump-config:
  dump-file-directory: /tmp/
```

## Test Plan 

Tested by simulating forks and replaying from dumps.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes