[chore] clippy to run with all features enabled
## Description 

clippy seems to be ignoring files when those depend on features that
aren't enabled when it runs. This is leading to build failures (ex
[this](https://github.com/GeunhwaJeong/haneul/actions/runs/4942463072/jobs/8836005749)
and [this](https://github.com/GeunhwaJeong/haneul/actions/runs/4910003494))
frequently as our rust CI is passing but when it comes to build CI it
fails and mostly goes undetected.

The PR is enabling `--all-features` so we align the rust & build CI. It
is also fixing some linter failures.


## Test Plan 

Did run locally

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes