periodic compaction of sst files
PR adds an option to enable periodic background compactions for the
objects table, ensuring that all SST files periodically go through the
compaction process. This feature helps address the issue where we
observe a large number of L4 SST files consisting mostly of range
deletes but not undergoing the compaction process due to internal
RocksDB heuristics.

This approach is intended to be used as a complement to aggressive
pruning and is disabled by default.
It is somewhat equivalent to RocksDB's `periodic_compaction_seconds`
setting, but the latter is not very flexible and can cause dramatic
resource consumption once enabled. The approach implemented in the PR
processes one SST file at a time