[DbTool] add a few commands to modify database with minimal safety checks
## Description 

There are needs to repair corrupted databases sometimes. Adding a few
commands to remove transaction and its outputs, and remove lock.

## Test Plan 

Tested on fullnodes.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes