[rpc][bugfix] Remove the usage of get_past_object_read from get_coin_metadata
## Description 

The current implementation of `get_coin_metadata` relies on
`get_past_object_read`, which stops working if the fullnode has pruning
enabled. This is unnecessary because we only need the latest version of
the coinmetada object(if it is wrapped, then it's okay to return null).

## Test Plan 

tested locally

```
curl --location 'https://mainnet.haneul.rpcpool.com' \
--header 'Content-Type: application/json' \
--data '{
    "jsonrpc":"2.0",
    "id":1,
    "method":"haneulx_getCoinMetadata",
    "params":["0x2::haneul::HANEUL"]
}'
```

```
{
    "jsonrpc": "2.0",
    "result": {
        "decimals": 9,
        "name": "Haneul",
        "symbol": "HANEUL",
        "description": "",
        "iconUrl": null,
        "id": "0x9258181f5ceac8dbffb7030890243caed69a9599d2886d957a9cb7656af3bdb3"
    },
    "id": 1
}
```

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Fixed a bug for `get_coin_metadata` which can return null when the
fullnode is pruned