use camelCase for `unwrappedThenDeleted`
## Description 

Previously, `effects` had an `unwrapped_then_deleted` field on ts-sdk.
This is an issue since jsonrpc returns the field as
`unwrappedThenDeleted`.

## Test Plan 

Locally built ts-sdk and explorer. Verified that visiting this page:
http://localhost:3000/txblock/3T2AYzJTvW5Hg4JhZaTgeFaTPSetyEBYZ44bVzScfibS?network=mainnet
did not result in any errors printed to console.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes