move metered-channel to haneullabs-metrics
## Description 

This PR moves metered-channel from narwhal to haneullabs-metrics and uses it
on `streamer` to monitor the channel size.

## Test Plan 

Tested it locally and see the metrics appeared.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Moves metered-channel from narwhal to haneullabs-metrics and uses it on
`streamer` to monitor the channel size.