set alt server name for anemo
## Description 

update anemo to
https://github.com/GeunhwaJeong/anemo/commit/1bfa7842f66a85d35345b5ada358df90c3efd579
which includes the `alternate_server_name` change. Set primary server
name as `haneul-{chain-id}`.

## Test Plan 

tested in labnet.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
update anemo to
https://github.com/GeunhwaJeong/anemo/commit/1bfa7842f66a85d35345b5ada358df90c3efd579
which includes the `alternate_server_name` change. Set primary server
name as `haneul-{chain-id}`. Note that if node starts to run this version
in testnet/mainnet before the update is officially rolled out, it's
likely the node won't be able to state sync.