fix big amount bug and 0 balance validation
## Description 

Describe the changes or additions included in this PR.

Fix the validation showing a 0 balance issue. Setting stale time to 5000
fixes it. useQuery queryFn
before


https://github.com/GeunhwaJeong/haneul/assets/126525197/41cd560d-03a5-4ce7-bfed-83374e6a94ea


after 


https://github.com/GeunhwaJeong/haneul/assets/126525197/058f6e17-46b2-4cc9-928e-6624a4442c06


## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes