Remove unnecessary workspace versions
## Description 

A few crates don't/shouldn't inherit workspace version. Give them
individual versions.
Also the open-json-rpc is generating version using haneul-core's version,
which is always the workspace version. This is wrong to me. This PR is a
short-term fix to just use workspace version directly in open-json-rpc,
but we should make it a real version.

## Test Plan 

cargo build

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes