Remove accounts from TestCluster
## Description 

TestCluster.accounts can be obtained from the wallet context within, and
hence it's redundant.
Since it's a testing method we don't particularly care much about
efficiency considering that the typical number of accounts in a cluster
is small. If we really want to cache it, we should cache inside wallet
context instead of the cluster.
Removing it from TestCluster.
This will set up the code better for consolidating TestCluster and
Swarm.

## Test Plan 

cargo test

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes