[wallet-ext] - enable transfer of ob kiosk items
## Description 

Enables transfer of items contained within an OriginByte kiosk. There's
some duplicate effort in finding the Kiosk for a given object, but
wanted to get this in so that transfers work.

## Test Plan 

Tested transfers to addresses both with + without existing kiosks

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes