Cancel submission waits after close of epoch
A node that is trying to send end-of-publish, but is waiting for locally
executed transactions to be sequenced, may end up waiting a long time if
the elected submitter of that cert has failed to submit. Instead, at
reconfig time, we should immediately submit everything we are waiting
for.

This greatly reduced reconfig pauses while testing my problematic PR
, and it may provide some benefit even during normal operation.
There were no obvious problems caused by the temporary increase in
amplification even though we were above 4K tps, so the inflight
submission limit appears to be doing the right thing.