[haneul-adapter] Move execution_mode and programmable_transaction/types
## Description

...to `haneul-types`, as another part of the interface between the
execution layer and its callers. The following changes were made to the
interface to make this work:

- `ExecutionMode` functions that previously accepted a `&mut
ExecutionContext` now accept a `&impl TypeTagResolver`, as that is all
this parameter was used for (for dev inspect transactions). This also
simplifies the constraints on generic type parameters in this module.
- `ObjectValue` no longer has constructors from `MoveObject` etc,
because it has been moved to `haneul-types` but these constructors are
implemented in terms of the execution layer. They have been replaced by
private free functions within `haneul-adapter`.

## Test Plan

```
$ cargo simtest
$ env HANEUL_SKIP_SIMTESTS=1 cargo nextest run
```