Simplify `handle_checkpoint_from_consensus`
## Description

Based on feedback, now this PR simplifies
`handle_checkpoint_from_consensus`. Previously there is a branch to
handle when the checkpoint is too new (newer than the
highest_verified_checkpoint). I think this case wouldn't happen because
consensus sends checkpoint in order. So we consolidate the two cases
into one, which is, whenever consensus sends a checkpoint, we do the
following things:
1. check the previous digest matches. This is to ensure state sync and
consensus do not brain-split.
2. if the checkpoint is older, ignore and exit
3. update metrics and bump watermark
4. notify others

--------------------------
# Old

## Description 

When consensus sends checkpoint y, it means all prior checkpoints are
verified and sent, so `highest_verified_checkpoint` can be safely
updated without looking at the checkpoint store. However it's a bit
tricky for `highest_synced_checkpoint`, because there are a few cfs
(e.g. `full_checkpoint_contents`) that could only be updated in state
sync checkpoint contents. If we bump this watermarks, those cfs may not
be filled at all.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Do not update highest_synced_checkpoint when consensus sends a too new
checkpoint