[transaction orchestrator] also subscribe to effects when waiting for a WIP transaction
## Description 

Some transaction timeout happens with a data race in this way (ordered
by time):
1. client sends Tx (1)
2. Tx(1) is processed, quorum driver notifies transaction orchestrator
the tx finishes.
3. transaction orchestrator gets the notification, working on finishing
up (e.g. removing the tx from the write ahead log, local execution)
4. client sends Tx(2) again, Tx(2) registers for notification, and see
the same tx is already WIP, so it awaits.
5. transaction orchestrator cleans up write ahead log and finishes local
execution, responds to Tx(1)
6. Tx(2) times out because no more notification is coming

This PR fixes it by also subscribing to effects when it awaits, such
that if the effects is available before the tickets, it means it misses
the notification, and will ask quorum driver to form a cert again.

This PR also removes `fn notify_read_executed_effects_digests` because
it's not in use.

## Test Plan 

This change is a bit hard to test because of the nature of race
conditions. But we make sure it does not break the existing tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Fixes a race condition that may very occasionally time out a transaction
request when it's submitted multiple times. Removes unused function
`notify_read_executed_effects_digests`.