Use steps to charge for gas instead of buckets
## Description 

Use a round up function instead of buckets to charge for gas

## Test Plan 

Tests passing, I am also going to run a node to see the gas charged

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Complete release note will be tracked in
https://github.com/GeunhwaJeong/haneul/pull/12429 because there is more than
one change coming and it makes more sense to track them all in one
place.

We are changing bucket computation for gas charges. 
Before this protocol version gas for computation was put into buckets in
order to group transactions that were roughly similar into the same
price/cost.
That, however, has created an inconsistency in the way charges operate
and the way people optimize for gas.
With this change in the gas model, charges for gas do not follow the
steep progression that was defined with buckets. Charges are rounded up
to the closest 1000 and that is all we do