[consensus handler] reorder txns by gas price
## Description 

This PR adds a step in consensus handler to reorder transactions in a
Narwhal commit by their gas prices. Right now the reordering is purely
based on gas price, which will be later changed to be based on object
hotness, etc, as well.

## Test Plan 

Added a test.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Previously transactions in a Narwhal commit were ordered as the sub-dag
is flattened in a depth-first traversal. Now we have added a round of
ordering that orders the user transactions in the same commit by gas
price, where a transaction with higher gas price will be added for
execution in the transaction manager first.