[haneul-execution] CutPlan::execute
## Description

Implement the process of executing a cut plan to copy packages and
fix-up manifest files (for the copied crates, and the workspace).

Also implements `CutPlan::rollback` which is used by `execute` to
clean-up after itself, if it fails. To support this, `CutPlan::discover`
has been updated to track the new directories it needed to create.

This PR does not address various usability issues (adding context to
error messages, a dry-run option, etc) which are to come.

## Test Plan

New unit tests:
```
haneul-execution/cut$ cargo nextest run
```

Try the CLI on a real example (warning, this will make changes to the
repo's working tree):

```
haneul$ cargo run --bin cut -- --feature v1                    \
  -d haneul-execution/latest:haneul-execution/v1:-latest          \
  -d external-crates/move:external-crates/move-execution/v1 \
  -p haneul-adapter-latest                                     \
  -p haneul-move-natives-latest                                \
  -p haneul-verifier-latest                                    \
  -p move-bytecode-verifier                                 \
  -p move-stdlib                                            \
  -p move-vm-runtime
```

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes