Stop depending on wrapped tombstone
## Description 

This PR simplifies the definition of unwrapped_then_delete in
transaction effects by introducing a new feature flag.
The flag is currently disabled.
With the flag is set, we no longer depend on tombstone when deciding
whether to include an object in unwrapped_then_delete. Since we no
longer query the store, we don't have the previous version number
anymore.
So we added a new variant in ObjectChange to cover UnwrapThenDelete,
instead of inside DeleteKind.
With this, we no longer need to depend on wrapped tombstones.
StateAccumulator can then also be simplified to ignore them.
At the epoch boundary where this is enabled, we also re-accumulate the
state and override the existing state.

## Test Plan 

Any suggestions on how I may test this?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes