[Narwhal] disable min_delay_timer after skipping rounds in proposer
## Description 

Currently `min_delay_timer` is reset in proposer for the next round,
after skipping rounds because of receiving higher round parents. The
network mostly increments rounds at an interval of `min_header_delay`.
So the following can happen,
- Primary A receives parents from round 15, skips proposing round 14.
- Primary A waits for min_header_delay=0.5s to propose at round 15.
- Primary A receives parents from round 16, skips proposing round 15.

This change avoids the additional wait in `min_delay_timer` when the
proposer skips rounds, to help the lagging primary become get in sync
with the network.

This change is intentionally simplified. A bigger refactor along with
some fixes to proposer will be done in a future PR.


## Test Plan 

This has shown to make header proposal rate more consistent with less
dips across validators in private testnet.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes