Disable fullnode local execution object fast path
## Description 

The logic is incompatible with TransactionManager AvailableObjectCache.
Populating objects in AvailableObjectCache is essential for correctness
in TransactionManager::enqueue(), so disabling AvailableObjectCache for
any or all objects requires refactoring TransactionManager. Also,
fullnode local execution itself is disabled now.

## Test Plan 

CI.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes