Increase reference gas price slightly to avoid failure after epoch change
## Description 

On epoch change, a transaction that will be retried might fail due to
having too low reference_gas_price set. This increases the gas_price
slightly. Today there is no other consequence, in the future this might
mean that the gas payed for the transaction will be slightly higher than
lowest possible.

## Test Plan 

Existing tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes