Use delete_range instead of clear() to avoid race condition
because clear() drops and re-creates the table without holding a lock,
it can cause crashes if other threads happen to try to write to or read
from the table being cleared. I don't think it makes sense to add
locking to every DBMap access just to fix this case, so I'm just going
to avoid calling clear().