[storage] Add per-epoch object marker table for shared object deletion and transfer-to-object
## Description 

This adds a new table to the ``PerpetualTables` that holds the set of
received and shared deleted objects during the epoch.

This table is not currently used anywhere and is unused. However, both
shared-object deletion  and transfer-to-object  PRs rely on
this table.

This table can be pruned at epoch boundaries so the table key contains
the epoch ID.

laura-makdah and I have discussed and we're happy to either land this
soon, or to wait until one or the other above PRs is ready to land and
land this change at the same time. Either way this change on its own
should be fine to land any time as the new table is not used anywhere
yet.

## Test Plan 

Not tested yet -- functionality will be tested in the two PRs being
built on top of this.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes