[Congestion] reduce per-object transaction queue length limit from 1000 to 200
## Description 

When transactions are congested on the same shared object, the cheapest
transaction types (e.g. incrementing a shared counter) support 50 ~ 100
transaction/sec per-object from testing. We need to reduce the
transaction queue length limit to a more reasonable value, to avoid
delaying checkpoint building and epoch change too much when an object is
congested.

It is still possible to cause congestion on an object with more
expensive transactions locking the object for > 0.1s. In future, a limit
will be added for the max queueing time, e.g. do not sign new
transactions touch a shared object when there is a transaction waiting
for the object for >= 2s.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes