[move-compiler] Added self transfer linter
## Description 

This implements a linter warning to be displayed when a function
transfers an object to the sender of the transaction instead of
returning it from a function.

It brings linter support based on the Move compiler to the same level as
the initial attempt to implement linters based on the Move model
(https://github.com/GeunhwaJeong/haneul/pull/11212)

## Test Plan 

A test has been added

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
When building Move code, additional linter warnings related to
transferring objects to the transaction sender may appear. The goal of
this linter is to encourage developers to return object from functions
rather than transferring them to the transaction sender in order to
increase composability of functions in programmable transaction blocks
by allowing callers to directly use the returned object.