Add HaneulAPIClient and HaneulHttpTransport classes
## Description 

This PR adds a new `haneullabs/haneul.js/api` export that exports a couple new
classes:

* HaneulAPIClient: This is a replacement for the concept of a Provider. It
servers as an API client for making RPC calls, and subscribing to
events/transactions.
* HaneulHTTPTransport: This wraps the RPC and websocket clients and
implements a new HaneulTransport interface used by the HaneulAPIClient which
defines a generic `request` and a `subscribe` method for making RPC
calls and subscribing to events.

The Transaction builder has been updated to accept either a `Provider`
or a `HaneulAPIClient`, and these classes should be interchangeable. The
only differences are in how they are exported and constructed.

## Test Plan 

Tests will be updated to use the new `HaneulAPIClient` everywhere in a
future PR

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes