[haneul-execution] Test to make sure generated lib.rs is not changed
## Description

Add a step to the rust `test` workflow to regenerate `lib.rs`, so that
if the PR has inadvertantly introduced a manual change to this module,
the `change-files` step will identify that and complain.

## Test Plan

Lints and formatters:

```
$ flake8 scripts/execution_layer.py
$ black scripts/execution_layer.py
```

Diff the generated (`--dry-run`) contents before and after generating
`lib.rs` for real. If there was a difference before, it should be gone
after generation runs for real:

```
$ diff -u ./haneul-execution/src/lib.rs \
  <(./scripts/execution_layer.py --dry-run generate-lib)
$ ./scripts/execution_layer.py generate-lib
$ diff -u ./haneul-execution/src/lib.rs \
  <(./scripts/execution_layer.py --dry-run generate-lib)
```

Check in CI that the step runs and catches a discrepency -- fix the
discrepency and notice CI finishes successfully.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes