[move-compiler] Keep running tests in presence of compiler warnings
## Description 

This fixes a problem when the test runner stops running tests in
presence of compiler warnings which is quite annoying for the
developers.

As an added bonus, with this PR we will no longer print compilation
results twice (also annoying).

## Test Plan 

All existing tests must pass.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

When running Move unit tests, the presence of warnings no longer
prevents tests from running and the compilation results are now printed
only once when executing the `haneul move test` command.