[Protocol Config] Mark get_for_max_version as unsafe
## Description 

See main comment in `crates/haneul-protocol-config/src/lib.rs` for reasons.
This is currently in use in places that it shouldn't be. Migration will
take more work, but for now, this should stop new incorrect usage.

## Test Plan 

`cargo build`

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes