Move more components to shared UI
## Description 

This migrates `Text` and `Heading` to the new UI base package. This also
completes the migration for `LoadingSpinner` (which I renamed to
`LoadingIndicator` to give us wiggle room on the UI, and can logically
be extended to other types of loading UIs).

Now that we have Text generalized, it's probably worth looking into
adding some of this into the wallet so that we could re-use some of the
base components there.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes