[rpc] logging macro logs original error instead of RpcError
## Description 

Currently, when we hit the `with_tracing!` macro, the error is already
an `RpcError`. Down the line, if/ when we want to replace internal
errors with a generic error string, we'll lose the debug information in
our error logs. This PR modifies the macro to log and map the original
error instead.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes