[wallet-ext] - Updated Kiosk UI / Enable Transfer for Haneul Kiosk
## Description 

This PR updates the UI for displaying Kiosks in the wallet. It also
includes updates to allow transfer of items within a Haneul Kiosk for
non-locked items. Also includes UI for representing assets that are
locked in a Kiosk



https://github.com/GeunhwaJeong/haneul/assets/122397493/e76b86a6-5648-4b90-ae2e-d4811072ba77




## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------