sort objects with HashMap
## Description 

1. remove fast path commit method. We came to the consensus that indexer
will not rely on fast path.
2. for object changes, instead of grouping and returning a nested
vector, using a HashMap and return the latest objects.
3. for mutated objects, apply chunks.
4. clean up experimental code


## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
1. remove fast path commit method. We came to the consensus that indexer
will not rely on fast path. 2. for object changes, instead of grouping
and returning a nested vector, using a HashMap and return the latest
objects. 3. clean up experimental code