Fix a bug in loaded child objects
## Description 

When executing advance epoch transaction, if we are also publishing a
new system package, we would be calling the execution engine twice, and
hence calling `save_loaded_child_objects` twice. We cannot just override
loaded_child_objects otherwise it will override the results of the first
execution.
In general, we cannot assume that we are only calling the execution
engine once.

## Test Plan 

Once we depend on this in , tests will cover it.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes