[ts-sdk] Add generator for RPC types
## Description 

This is a fairly large PR. I re-structured the commits so that they can
be viewed individually in order in a way that should make this easier to
review.

There are a bunch of places in explorer and wallet where we needed to
add some more explicit types. The new generated types use unknown or a
specific type in a few places that previously used any. This allowed us
to write a lot of things without adding extra type-hints or checks, but
the added safety of using unknown or accurate types should be worthwhile
in the long run.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes