[narwhal] align low scoring detection on submission with leader schedule
## Description 

The changes on the new leader schedule algorithm will always exclude the
X% by stake of validators with the worst scores. Practically we don't
have good reason to continue submitting transactions to those validators
to sequence in consensus as they'll never be leaders hence they have
slimmer chances of sequencing transactions.

It makes sense to align the low scoring detection on the submission side
with the logic of the leader election schedule. Thus we'll make sure
that only the validators who participate on the latest (updated)
schedule will receive transactions to sequence them. For that reason we
are removing the `MED` approach and we go with a flat exclusion always
of the validators with the lowest scores and up to the designated
`consensus_bad_nodes_stake_threshold` . For instance, if the
`consensus_bad_nodes_stake_threshold = 0.2` then the nodes with the
lowest scores and accumulated up to 20% of stake will be flagged as low
scorers.

**Pros:**
* We make sure that we don't send transactions to any validators that
are not part of the leader schedule committee, thus we maximize the
changes of getting transactions sequenced in time
* Although MED is a great technique, it's hard to tune under different
network properties and harder to reasoning. Now things are more
straightforward and predictable

**Cons**
* The algorithm will always account "low scorers" - up to the dictated
fraction - even if we had all validators with perfect scores. This is
more of a concern for small environments (a few nodes) rather than in
environments like `testnet`or `mainnet` where we do have even the
current MED approach flagging a good amount of nodes as low scorers
* To tackle the above we could introduce some manual threshold, but
needs more experimentation/iteration

This PR is also 
1. cleaning up the
`update_low_scoring_authorities_with_previous_disable_mechanism` method
which should be now completely unused.
2. Shoehorning a test for long period of asynchrony for the new leader
election schedule algorithm

**Note:** that doesn't mean that a low scorer node will not be able to
sequence any transaction, for example the checkpoint signatures or the
capability notification messages will still be able to get sequenced.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes