[wallet-standard] Sign personal message
## Description 

This is a change to the Wallet Standard, which deprecates the
`signMessage`, and introduces a new feature, `signPersonalMessage`,
which is intended as a 1:1 replacement (which will always be in the new
signature format). This matches the current terminology in the updated
Haneul TS SDK, and should help remove confusion that has come up around
what is being signed and verified in the SDK and wallets.

I added `TransactionBlock#sign` in order to make a smoother transaction
from `signer.signTransactionBlock`.

While I was here, I reworked the burner adapter to use the new methods.