accounts refactoring
## Description 

Enables storing public data (address, public key, etc) for accounts to
allow showing account data even when the accounts are locked. (this only
runs through the test `#/accounts-dev` page for dev mode and when
`NEW_ACCOUNTS_ENABLED` constant is set to true)

* adds new Account classes that handle specific account operations based
on the account type
* also adds AccountSources that mainly used to store shared data between
same type of accounts. eg. MnemonicAccountSource that stores the
mnemonic and can be shared across the mnemonic accounts
* accounts now can be read unlocked (unable to do write operations) but
able to show balances etc. This will be mainly a UI privacy feature
where we check the `lastUnlockedOn` to be less than max allowed
* store new accounts using indexedDB with dexie js
    * to address any concerns of browser clearing the indexedDB data
      * added `unlimitedStorage` permission
* and a mechanism to import and export the db from local storage. After
adding a new account or account source indexedDB is being saved to
`local storage` and when the database opens we check to import the
exported data
* integrate Sentry in the service worker context, to allow logging
errors about indexedDB being cleared from the browser/user
* added some test pages to test adding/removing accounts that will be
deleted later
* each account and account source unlocked individually and stores any
sensitive data in session storage encrypted
* currently accounts and account sources get locked only manually (or
when browser clears the session storage) - will revisit this later when
we have a better idea about locking accounts



https://github.com/GeunhwaJeong/haneul/assets/10210143/f0a07233-563b-4b41-b6f8-ebab3d17abca



https://github.com/GeunhwaJeong/haneul/assets/10210143/eed04ad2-2a23-4ca7-abf6-50a0b9880864



https://github.com/GeunhwaJeong/haneul/assets/10210143/a7a776ab-d66d-4f84-b418-b1925c15b290



closes APPS-1099

## Test Plan 

change `NEW_ACCOUNTS_ENABLED` to true and visit `#/accounts-dev` page

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes