[move-compiler] Fixed abstract interpreter bug
## Description 

- Fixed abs int bug that was fixed in the bytecode verifier, but not the
compiler
- Addressing https://github.com/move-language/move/issues/1068

## Test Plan 

- Added a test

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Fixed an edge-case bug with the Move compiler that could result in
control-flow sensitive errors not being caught before reaching the
bytecode verifier.