Deprecate wallet adapters
## Description 

This deprecates all wallet adapters, in favor of the wallet standard
being _the_ way to interact with wallets. Wallets that do not conform to
the standard are expected to bridge to it, instead of the current
approach of using adapters to adapt over wallets. This dramatically
simplifies a lot of logic, should be considerably less error-prone,
makes it easier to extend functionality in a consistent way (using the
standard instead of adapters), removes several packages, etc.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes