[dapp-kit] Add ability to configure networks to HaneulClientProvider
## Description 

This is getting into react stuff where I am not very confident I know
what I am doing. The goal here is to provide a Context provider that
supports switching between networks, and allows providing either
pre-configured HaneulClients for each network, or a config object (with
customizable options) that can be used to create a client.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes